auth logic *almost* complete
