+++ element/user_count.txt
@@ -1,4 +1,4 @@
-User	1744
+User	1747
 Plot_Generated	23846
 Spectra_processed	0
 Spectra_Processed	418232
